Initial Commit
Game was made in the span of 12 hours, and used for a birthday party where birthday boy got to choose an extra feature and watch it added live in front of them. Added feature was loot crate boxes for AI helpers. It was a little buggy, but otherwise worked and the kids loved it! (This game was made back in April of 2018)